autorelease_for_round(), in the "emergency valve" code:
Formerly, we would simply use the first project in the query result.
However, this ignores two (admittedly unlikely) possibilities:
(a) the query might yield an empty result set (in which case,
    trying to fetch the first result would raise an error); or
(b) the first project might fail to release (in which case,
    there's no error, but we miss the chance that some later
    project might succeed).

So instead, loop over the results of the query (if any),
and break out of the loop as soon as a release succeeds.