Introduce $where_clause for the text of the WHERE clause passed to select_from_site_past_tallies_and_goals().
For the 'all_time' case, just use $where_clause = '',
because we want all the available data.

For the 'curr_month' and 'prev_month' cases,
we used to say:
    WHERE $start_timestamp < timestamp AND timestamp <= $end_timestamp
in which the start and end timestamps are calculated
for the desired month in the server's time-zone.

But this over-precision is not robust to a change in server time-zone
(which pgdp.net recently experienced).

So instead, say:
    WHERE {year_month} = '$year_month'
and let select_from_site_past_tallies_and_goals()
shoulder the responsibility for handling a time-zone change
(which it doesn't yet, but will soon).