Change the date format from "n/j/Y" to "Y-m-d".
For one thing, jpgraph renders the date vertically,
with the "left" of the text on the bottom, "right" on top.
With "n/j/Y" format, this puts the year closest to the axis,
but the year changes slowest, so tends to be the least informative.
With the "Y-m-d" format, the day of the month is closest to the axis.

For another, "n/j/Y" is a fairly US-centric format,
and one that we don't use much on the site.
"Y-m-d" is more international, and maybe
less prone to misunderstanding.

Also, "Y-m-d" matches the format in the "graph generated at" line
in the bottom-right corner.