Move generated files to .template files and add them to .gitignore.
SETUP/configure changes files tracked in git. This increases the chances
that a developer will accidentally check in those changes which is both
a security risk and messes up the source code.

Instead, lets move all files that will be configured to have a .template
suffix and have SETUP/configure operate just on those. Then we add
the generated files to .gitignore such that they will never be checked in.