Create forum account before the DP account.
Create the forum account before we create the DP account and remove
the account activation. If the forum account creation failed, stop
and return a message on how to get help. Note that the phpBB3
error code is included as an HTML comment in the returned page
to allow dphelp to access the activation URL themself and see the
error message.