page_summary_sql(): Remove superfluous join on phpbb_users ...
(A long time ago, we used it to get the user's forum id, to link to their
forum profile. But since 2004-09-08, we've linked to their DP profile.)

This actually fixes a bug on pgdp.net, because the 'live' phpbb_users table
is no longer in the default database, so the phpbb_users table that the query
was joining with didn't have records for new registrants, and so (because
it's an inner join) those users didn't show up in the query's results.