Pre-process username/password before doing auth.
phpBB3 does pre-process/escaping of username and passwords before
sending them to auth->login(). If we don't do similar processing
then users with HTML special characters (like &) in their passwords
can log into the forums, but not into DP. This shouldn't be a problem
for usernames since DP restricts what characters can be in a username
to be more strict than phpBB but we'll do it for consistency anyway.