Encode password before hashing it.
phpBB3 does some encoding of fields after they are POSTed in from the
web browser. This includes the user password before it is hashed and
stored in the database. In order for user registration to work with
passwords containing characters that are modified by the encoding, we
have to do the encoding too before hashing it and storing it in the
non_activated_users table.