Wire up the project_upload.php script
Provide a link to the project_upload.php script in the same place
where we show FTP reminder information.

Update the configuration.sh script to provide details on the
$_UPLOADS_* variables and how they are used by the new script
and if they are using FTP.

Update installation.txt to discuss creating the uploads directory
and optionally the FTP/scp user if so desired.