Use common error message to prevent data leakage
To prevent leaking information about the local filesystem, all error
messages about if a file/directory exists or not need to be the same.
If we gave one message if the file exists and another if it wasn't
in a normalized form, we allow them a programatic way of determining
information about what files/directories exist on the system.