Set default values of last_modified columns
At some point on pgdp.net, we updated the default value of the
last_modified column for the authors and biographies tables to be
the current timestamp. The column is set to set this column to
the current timestamp upon update as well.

Given that these columns have existed since the tables were introduced
in 06, but are never set anywhere in the codebase, having the DB
maintain the value of this field seems logical.