Don't strip out DEFAULT CHARSET by default
We really want to create the tables using the specified charset
(latin1). This is particularly important given that we give
instructions in installation.txt to create the database with UTF8
and tables will inherit this unless told otherwise. Since we still
support MySQL 4.x, we leave this charset-stripping in, but disable it.