Be more restrictive about messages output to stdout
These files are expected to be run via cron. On most systems, cron
captures command stdout and includes it in an email to the system
user. With this in mind, we should only write error messages, or
messages that we really want the system admin to know about, to stdout.

Now that they are (somewhat) cleaned up, ensure that the output is
being captured instead of being redirected to /dev/null.

README.txt encapsulates information for system admins as to how these
scripts are expected to be used as well as best practices for
developers when implementing them.