Reorganize and improve configuration documentation
The configuration.sh script is one of the most important pieces for
site installation and yet it's rather haphazard in its organization
and documentation. This commit is a massive upheaval in an attempt
to provide better organization and to flush out documentation across
the board.

Of note, the code also attempts to automatically detect binaries
for executables where possible instead of requiring the user to set
them. It also attempts to use saner default values where possible.