Script to help update configuration.sh on upgrade
configuration.sh is likely to change from release to release, but most
of the parameters will be the same, yet must be transferred to the new
configuration.sh upon upgrade by hand. To reduce the pain in migrating
the parameter values from an old file to a new one, include a script to
assist with this process.