Make jpgraph 3.5.0b1 show nice legends too
We currently support both jpgraph 1.27.1 and 3.5.0b1. In order to
make the latter show nice legends, we need to SetColumns(1) on it.
According to the documentation this is suppose to be the default
but it doesn't seem to be working as documented.

While we're here, I'm adding a small drop-shadow on the legend
to better separate it out from the graph, as well as moving the
legend around a bit to make it less obtrusive over the data.