Only update t_last_activity once every 5 minutes
Currently, for every single page load we update the t_last_activity
column in the users table for that user. There is no need for this
level of logging and it results in thrashing the users table for
negligible gain. Instead just update the table no more than once
every 5 minutes.