Only show installed languages in drop-down
Given that only unauthenticated users will see the drop-down, only show
installed locale translations (translators must be logged in to help
with translations). en_EN is a special exception since that's the native
format of the strings in the code.

Because we include all enabled translations and the javascript only
triggers on change, we need to include a no-op value in case only one
translation is available.