Redirect authenticated user to Activity Hub
This fixes an annoying corner case in user activation that makes for
a subpar experience. If after using the activation URL, the user uses
the navbar to sign in, they are redirected *back* to the activation
URL and presented with inaccurate text.

This change checks to see if the user is activated and the activated
user is logged in, and if so redirects them to the Activity Hub.

In addition, if a user creation error occurrs, log the error to the
filesystem in addition to outputting it in the HTML comments.