Move round_backlog graphs from noncvs into git
The round backlog graphs on pgdp.net are used within news items to
draw attention to rounds that need the most help based on the amount
of work in them. Move these graphs into git so that their strings
are included for localization and for other DP sites to use.