Automatically call slim_footer() upon shutdown
Like we do for the header/footer in theme.inc, register slim_footer()
as a shutdown function within slim_header(). This ensures that whenever
the script finishes we close down the HTML tags. It also means that
we don't ever need to call slim_footer() explicitly.

Because slim_header() is used for outputting framesets, which don't
include a <body> tag, we need to ensure slim_footer() doesn't output
a </body> tag if one hasn't been opened.