Allow passing projectid without projectID prefix
Having projectID prefix all of our project IDs is quite silly. While not
an issue for internal links, it would be more elegant for external links
to use the 13-digit hex number. We can leverage the validate_projectID
function to accept both forms and return the long one.

Note that we still use the full form internally, this is just a
convienence for external URLs.