Don't treat translatable strings as fixed values
The dp_proof.js file uses the 'Default Browser' string as a special
value. In other places we also compare against BROWSER_DEFAULT_STR.
This is a bad practice since the string may change, either in the
code itself or as a translation.

Instead, treat index 0 with an empty value in the font and size
arrays as a special case meaning the browser default.

Also:
* Change $f_f and $f_s to reasonable names:
    - $f_f = $proofreading_font_faces
    - $f_s = $proofreading_font_sizes
  and make them associative arrays. This retains the requirement that
  the keys remain constant but lets us place them in an arbitrary order.
* Standardize on 'Browser default' as the preferred case site-wide.
* Change to using YYMMDD## as the JS versioning method since we're
  no longer in CVS and git version hashes are way too long