Make Usertest.php pass PHP lint test
This file has never been completely implemented, as shown by the
mismatched function names between it and the (never used) class it
purports to implemented. Moreover, it doesn't look like it ever
passed a PHP lint test.

Still, to ensure all PHP files in the repo pass a lint test, fix
the two conditions where it didn't:
* empty() prior to PHP 5.5 only supports variables.
* mis-matched set of parens.

This still isn't completely implemented and is unlikely to work
at all.