Bugfix: Set variable $project_is_bad.
This fixes a bug I introduced in CVS rev 1.86 (2006-06-20).
Formerly, with $project_is_bad undefined, it was always effectively false,
so if a bad page report caused the *project* to be marked bad,
this script would continue as if the project were still okay,
which, if the user had picked "Continue Proofreading",
would result in an odd error message.

Now, when the project is marked bad,
proofreading stops and the user is directed back to the Activity Hub,
regardless of whether they had picked "Continue Proofreading".