Move udb_user.php inclusion to local context
We use the variables defined in udb_user.php in very few places.
Instead of loading them in the global namespace, just load them
in the few places we need them. Of note, this makes them local
to the DB connection function.

The impetus for this change is the upcoming inclusion of phpunit
tests. As-is the tests will be unable to connect to the database
because the udb_user.php variables are only implicitly global,
not explicitly global. That is, if udb_user.php is included from
a global context they are global and accessible to dbConnect(),
but phpunit calls bootstrap and test code outside of the global
context. I could have just defined the variables in udb_user.php
as explicitly global to solve this problem, but there's really no
need for them to be global in the first place.