Restructuring and expanding stylesheets
There were several issues with the initial structure of the new
stylesheets. default.php.css would not load on the server, and the
styles contained in it were applicable to any page with a statsbar.
Also, sans-serif font styling is dependent on the used theme, and
therefore it needed to be declared in theme.less. New styles for
recent HTML changes were also added.