Remove translations in new PO files
When creating the PO file from the POT file, msginit populates the
new translations with the original strings. This makes finding
untranslated strings in PO editors difficult and messes up the
count of translated strings in the Translation Center.

Now, after using msginit to create the new PO file, we post-process
the file to remove all translations.