Split authentication from user UI in login.php
Split out user authentication and session set up (which requires no
UI) from the UI portion reporting a login failure. This is part two in
reducing the possibility of code collisions between phpBB and the DP
code.

To pass the error from login.php, pass a URL-friendly key to
login_failure.php which is converted to a user-friendly error message
and displayed to the user. We also handle the case where no error code,
or an unrecognized error code, was indicated.