Don't load User.inc in login.php
DP and phpBB both have a User class. If we include pinc/User.inc
here, when we make the phpBB call to log the user in and the phpBB
code is loaded, we get an error that the User class has already been
defined.

Ideally, all user input would be validated before being passed down
the stack and used. In this case, however, that additional work isn't
necessary. Our phpBB2 login implementation correctly escapes $username
before querying against the table and phpBB3 does as well.