Ensure pages include misc.inc where necessary
Now that undo_all_magic_quotes() is in base.inc, several functions
don't need to include misc.inc so remove it. Also, there are several
files that use functions in misc.inc but that don't include it,
relying on something else to pull it in instead, which is bad.

The "who needs misc.inc" analysis was done via script grep'ing through
the code base determining which misc.inc functions were used where.