Update SettingsClass to PHP 5-style
Explicitly define the scope of class varaibles and methods. Move
ExecSQL (which IMHO is of dubious value) into the class to reduce
global function namespace pollution / collision. And if there are
errors, throw an exception instead of echoing them to the screen
(which might never get seen).

Throw an exception when trying to set an unsettable value instead
of dying so it is catchable.