Allow Settings to have multiple values
In some places in the code, we (ab)use the Settings table not as
a 1:1 key:value storage, but as a 1:many. Instead of special-casing
those instances, make the Settings table multi-value aware.

This retains the additional 1:1 access methods and adds three more
for the 1:many: get_values() / add_value() / remove_value().

Under the covers, we assume that all settings can be multi-valued
and store them all as arrays. We enforce the 1:1 uniqueness in
get_value().