Expand slim_header()'s capabilities
slim_header() is used for non-themed pages in three primay ways:
  1) "normal" pages with no special CSS or JS
  2) pages with special CSS or JS
  3) frameset pages

The last two are supported via parameters that toggle how the <head>
and <body> tags are manipulated, to allow the caller to handle it.

Instead of providing an unwieldy interface to toggle if certain tags
are output or not, we make slim_header() capable of handling multiple
different scenarios, such as:
  * custom CSS (via css_data)
  * custom JS (via js_data)
  * custom JS include files (via js_files)
  * custom tags in <head> (via head_data)

In addition introduce slim_header_frameset() for use when outputting
framesets.