Convert MARCRecord functions to a class
Convert the MARCRecord functions into a real class. Note that
pulling error_reporting(0) out of the file identified notices in
both it and editproject.php that were fixed.

In order for the unit tests to pass, we need to explicitly
declare $lang_data global in iso_lang_list.inc. This is necessary
because of how phpunit loads files (ie: not in a global context).