Remove dep on magic quotes from tasks.php
Removing the dependency on magic quotes from tasks.php involves
first cleaning up the double-encoded values in the database. This
is done via the 09/20160308_undo_task_double_escaping.php script.
Then we simply do much less processing in tasks.php when storing
the data and a little more when displaying it.