Create and use a themed table with CSS
Create two new table classes: themed and striped. Themed tables use the
navbar color and sidebar colors as their primary palette. The striped
class adds alternating sidebar/page background row colors to make the
contents easier to read.

In the stats code, remove all of the hardcoded references to $theme
and use the new table classes instead. This significantly simplifies
the code required to output the tables and the size of the pages too.