Create and use html_safe() function
PHP 5.4 changes how htmlspecialchars() works by assuming that the
string's character set is UTF-8. If a string that isn't UTF-8 is
passed in, an empty string is returned. This was further changed in
PHP 5.6 to use the php.ini default_charset if none is explicitly
passed into htmlspecialchars().

In order to work across all PHP versions without having to access
the $charset global variable all over the codebase, create an
html_safe() function to encapsulate the call to htmlspecialchars().