Unify id for mid-level page container
The CSS and HTML code had different names for the mid-level page
body container that provided the "display: table-row;" CSS rule.
As a result, this rules was not in effect, causing issues with
statsbar display width. This commit chooses one name for both
the CSS and HTML code, which should fix the issue.