Update quiz feedback message
When working on the current proofing or formatting quizzes, the user
is directed to the proper forum to post feedback messages and ask for
help, but many users end up clicking the "new topic" button once
they're there, and their question ends up as a new topic in the
Documentation forum rather than ending up in the proper quiz
discussion topic. This change updates the quiz response so it
indicates that the user wanting to provide feedback should visit the
forum and push the "post reply" button. (Change is installed in the
wfarrell sandbox.)