Task 1655: Add working links to Special Days Info
On the Special Days Info page (reached from the PM page) the result
column "Name" shows a link which one might assume would be a
working link that would issue a search for projects associated with
that special day. However, the link is merely for show, linking
back to the current page to demonstrate how the color appears when
a table cell on a round page contains a link.

This commit makes that column of the listing more useful by making
it a link to a project search for that special day.