Disable proofing interface assert
The assert is technically correct -- text_frame_std.php should only
be loaded if the proofer's proofing interface is set to standard.
Indeed, that is the only case for site-generated URLs, which was
nominally what this was meant to check when the enhanced interface
was added.

However, foofAid is loading the standard interface URL even if the
user's proofing interface is set to enhanced. This has no negative
impacts on the page being proofed, but generates noisy PHP errors
with no actionable item.