Change to use new DP logo
The new DP logo is a reimplementation of the old one into a vector
format for creation of raster images of arbitrary size. The new raster
logo is a PNG file supporting alpha transparency, instead of the
GIF format which does't. This means we don't need separate logo files
for each theme, further simplifying the theming (which will eventually
be replaced entirely with CSS files).

Original SVG files can be found at:
    http://www.pgdp.net/noncvs/dp_branding.php