Pull logo out of .css files
Now that we have a common logo we can use for all themes, pull
the logo out of the .css files and back into theme.inc. This allows
proper alt text to be placed on the image, as well as allowing
individual sites to have their own image instead of always pulling
from pgdp.net.