Create function to return translated graphic
For images with text in them, we need to return an image with
the text translated into the right language. If that's not available
we should return the translated text and show that.

This introduces the $dyn_dir/code_images directory that stores
all images referenced in the code in one place. Each subdirectory
under $dyn_dir/ should have one directory under it per locale.
For instance: $dyn_dir/code_images/news_header/en_EN