Move shortcut icon after mobile icons for FF
Firefox loads all rel="icon" and rel="shortcut icon" images and
uses the last one as the favicon.

We intentionally use white backgrounds on the mobile icons but a
transparent background on the favicon. If we specify the mobile
icons last, FF will use that as the favicon, which isn't what
we want. Fix this by swapping the order around and have the
favicon be last.

See bug: https://bugzilla.mozilla.org/show_bug.cgi?id=751712