Reduce num of globals and move to top of function
There's a bug where we use $projectid in an if() branch within
list_images() but that variable isn't available because its global
declaration is in another if() branch.

Reduce the total number of globals used by passing them into the
function and move all the global declarations to the top of the
functions.