Don't throw a RuntimeException on multiple values
Ideally we'd throw an exception if we detect that the number of
values is more than 1. The problem is that in rare cases two
processes can race clearing and setting values. Instead, we
resort back to what we did prior to supporting multiple values,
which is to just return one of them. The next time this setting
is set with set_value(), it should be reduced back down to 1.