Redirect guidelines to documents in the wiki
pgdp.net has moved the proofreading and formatting guidelines into
the wiki. Ideally there would be a robust configuration capability
in the code to allow for redirection of arbitrary faq/ pages, but
the current configure system makes using advanced data structures
like arrays difficult.

In the meantime put in this hack to redirect users who access the
proofreading and formatting guidelines on pgdp.net and pgdp.org to
the relevant pages in the wiki.