Throw exceptions instead of dying for validation
die()ing is a terrible thing to do inside a function as it prevents
the function from being used and tested elsewhere. The "right" way
is to throw an exception and give the caller a chance to catch the
exception first. If uncaught the end-result is the same -- the script
dies with the desired message -- but it plays better with others.

This allows things like the creation of unit tests to validate the
functions.