Introduce get_float_param()
To allow the validation of floating point numbers, we leverage all
the logic in get_integer_param() by renaming it to get_numeric_param(),
extending it to support floats, and creating lightweight versions of
get_integer_param() and get_float_param() that call it.

Update the tests to test floating point numbers as well.