Changes to check for following blank line
Message changed to say "Footnote"(etc.) (rather than "this") should be
followed by a blank line, clarifying what the problem is in case a
false possible issue is caused by an unmatched closing bracket in
the text. Placeholder %s used for this, and wording changed.
Thought break should be followed by a blank line" is now an issue so
that the tag is not coloured causing a markup conflict