Fix ancient bug re marking projects bad.
When pgdp.net first introduced bad pages (back in the R1/R2 days),
the two rounds had quite different criteria for
how bad pages caused a project to be marked bad.
Over time, the criteria became more alike,
but pages_indicate_bad_project() retained an if-stmt
to add a special case for R2 (round_number == 2),
such that a project was deemed bad if it had
at least one bad page and no available pages.

When pgdp.net changed to 4 rounds in June 2005,
we should have done something to eliminate the R2-specificity
of this code, but didn't. So it became a bug that lay unnoticed
until a week or so ago, when Walt Farrell spotted it,
after its effect in P2 was observed in projectID57affd75eab72.

There are various ways one could fix the bug
(i.e., eliminate the 'round 2' specificity),
but the least-impact approach is probably
to simply delete the special case code,
which is what this commit does.