Decrease Project constructor overhead
Creating a Project object should be fairly lightweight, but creating
project credits has a noticeable amount of overhead due to querying
the image_source table. Similarly until it's needed, there's no point
in querying the disk to see if the project directory exists or
querying the DB to see if the project table exists.

Move all of these object properties out of the constructor and into
a __get() so that we only incur their cost when they're needed.