Don't decrement n_available_pages < 0
In weird error conditions, usually from manipulating project tables
directly, the n_available_pages can be 0 when there are still pages
available. When this happens using MySQL 5.5 and we try to decrement
the counter, mysql throws an error as the field is unsigned.

Do some additional checking to ensure n_available_pages does not get
set to a negative value.