Pennsylvania Dutch as a project language
Add Pennsylvania Dutch as a recognized language for projects
at DP, as requested by one of our Project Managers. For ISO language
code purposes we use pdc as the 3-character code (based on some
versions of iso-639-3 showing it) and pd as the 2-character code
(not currently used in the ISO tables). For the name in the local
language we use Pennsilfaanisch Deitsch (as suggested by the
Project Manager).