Add patch to include PM contents in phpBB emails
There is no Prime Notify extension for phpBB 3.2 that will include
the PM contents in the forum notification email. Add a patch file
that will add that functionality. [checked in on behalf of wfarrell]

Update phpbb3 conversion documentation to include reference to this
file and how to use both phpbb3 patch files.