Add context-sensitive gettext function
Some strings need to be translated differently based on their
context. For instance 'PM' means both private message and project
manager and will need to be translated differently in French.
gettext supports this concept through the pgettext function, but
there is no equivilent in PHP. Thankfully, someone has already
solved this for us on the internet.