Make slim_header() use html_header()
We want styling for pages that use slim_header too, so simplify
everything by having slim_header() use the same start-the-page
function as the output_header() function does for themed pages.

This will have some small but noticeable impacts elsewhere that
we will probably need to mitigate, but it's the right thing overall.