Use the CSS/JS browser cache fixer for all files
Previously we were relying on a developer who changed certain JS
or CSS files to manually update parts of the code to essentially
force a browser to reread the file and pick up the change. This
moves that functionality to theme.inc and makes it valid for all
files. This will help tremendously during CSS development.