Merge review_work_instrumented3 into review_work
Back in at least 2011, possibly earlier, work was done to improve
the performance of the review_work.php script and have it put less
load on the database. That work cumulated in a noncvs script called
review_work_instrumented3.php. This merges that work into the
review_work.php script.

Note that the new version contains both a new query against the
user_project_info table and the old one against page_events. The
latter is still necessary to order the results in a way that is
useful to evaluators. We restrict the older query to just admins,
PFs, and evaluators to mitigate load on the system.