Escape page URLs upon consumption
Commit 195066a9 incorrectly escaped page URLs within the calling
function instead of where they were being used. This is a problem
if the URLs are used outside of HTML in header redirects or other
places where &s should not be escaped. Do escaping based on how
the resulting URLs will be used.

Most notably this fixes a problem where
"Save as Done and Proofread Next Page" from within WordCheck
resulted in an error instad of the next page.