Fix DP error in displaying a user's Google+ link
phpBB32 allows a user to add a Google+ link in his forum profile. DP
will display that info on the user's stats page, but incorrectly
adds a + before the user's Google+ profile number in the link, which
makes the link invalid. This change removes the + sign, allowing the
link to work.