Fix DP issue displaying links to a user's Google+ account
PhpBB32 allows a user to add a Google+ link in their forum profile.
DP will display that info on the user's stats page, but incorrectly
adds a + before the Google+ profile in all cases. The + should not
be used for old-style profile info (all numeric) but is required for
users with a custom Google+ URL (non-numeric).