Enable resumable, large-file uploads
Using resumable.js, a 3rd party JS library, we now allow uploading
of file sizes much larger than can be supported in a single POST
request by PHP. resumable.js does this by uploading the file in
pieces and then having the server reassemble them. If an upload
sessions is interrupted, uploading the file again will only upload
the pieces that did not make it sucessfully to the server.

resumable.js requires HTML 5 support and we only enable this method
if their browser will support it.