Access setting names in usersettings vary slightly
When we grant or revoke access to an activity we set a value in
the usersettings table as well as log the event to the access_log.
For historical reasons, Stage activities (proofreading rounds, etc)
include an .access suffix in the usersettings table (eg: P3.access)
whereas other activities do not (eg: manager). The $use_access_suffix
allows the caller to specify if the suffix should be included
without forcing the logic of what needs it into this class.