Update PHP and phpBB version support
After ea40ee5d merged we know the code works with PHP 7.0.

When we merged ba74ae04, we broke support for phpBB 3.0. This
is because accounts/login.php pulls in metarefresh.inc which
pulls in slim_header.inc. ba74ae04 changed slim_header.inc to
bring in theme.inc which eventually brings in user.inc which
breaks phpBB 3.0 because it also defines a User class. phpBB 3.0
is out of support and phpBB 3.1 was released in Oct 2014 so
rather than trying to hack a fix in, lets just say 3.1 is the
minimum supported version.