Remove success interstitial
After a successful action, it is far more streamlined if we simply
returned to the showdir function, outputting any success message,
rather than have an interstitial whose sole purpose is to report
the success message.

Uploads are treated separately since after the upload we do processing
on the file and give status on the processing as it happens.