Restrict nameofwork and author inputs to 255 chars
The size of the db columns for nameofwork and author is 255. Prior
versions of MySQL would silently truncate longer strings to the
size of the column, but modern versions throw an error. Prevent
users from inputting strings longer than 255 for these columns.

Without a doubt there are other instances where we should be doing
this type of enforcement, but this addresses this instance.