Fix background for highlighted cell
Cells with highlighted background color were being overridden if
they were on even rows because of the table.striped row CSS.
Change that to a style attribute to correct this.

While we're here, fix the number cells to be right-aligned to match
with our style guide and output proper attribute names for the number
columns.