Make User->load() private
To reduce the possibility of bugs, make User->load() private. This
ensures that an object can only be loaded by the constructor (or
the two new static functions), and prevent the case where a developer
has created a loaded object and then loaded different data into part,
but not all, of it.