Use type=number inputs for numeric form fields
For a better user experience, use number field types in forms where the
input is only numeric. This isn't perfect as some browsers will allow
the user to input an 'e' (expotenetial notation) and floating point
numbers are also allowed, but it's a step up from the wide-open
anything-goes that it was prior.