Do SR completion notifications at end of the hour
SR completion notifications run once an hour and send notifications
for any projects that finish SR within that hour, even if the
projects won't finish SR until the end of the hour after the job
runs. To minimize the chances that the notification will be sent
out before the project actually leaves SR, do the notifications
at the end, rather than the beginning, of the hour.