Add a "text" mode to Format Preview (#145)
This adds a "text" mode where inline markup (i, b, f g,u) is replaced
by a single character to show how text will appear in the "plain text"
version of the book to get correct alignment of tables etc. Small caps
markup is not marked in this way. "Wrap" is now a checkbox so that the
Text mode can be used with or without re-wrapping.