Add codepoint validator (#154)
Add a validation check before every page save to ensure that all page
characters are valid for a project. This prevents the user from being
surprised when the backend validation removes invalid characters. If
all characters are valid the check is not shown and the page is saved
as usual.