Add a buuton to remove bad characters.
Suppose the user insert a combining diacritical mark. The text will
be normalised. 3 cases:
1, a good character resulst -- ok
2, a bad character results and will be marked.
3,  no more normal form exists and the diacritical itsef is marked
This is invisible so the user may not know what the bad character is.
This feature allows him to proceed.
Interestingly in php regexes \X matches a composed glyph but this
is not available in JS regex.
It would be possible to do this functionally but seems a lot of
work for little return.