Update Project Quick Check to support UTF-8
All of the characters in the database should be UTF-8 encoded now, so
many of the character checks are no loner valid. A few, however, may be
useful so be sure and retain them:
    * HTML entities
    * double-encoded UTF-8

In addition, add a check for invalid project characters in the text.
While these should not be added to new projects, they may continue to
exist for older projects for a while.