Drop redundant indexes
page_events is the largest table on pgdp.net and contains indexes
that are no longer used but are taking up disk space: the index size
is 2x the size of the data itself.

Drop the unused indexes on this table that aren't used in queries
as confirmed by explain plans.