Add an editable site-wide alert bar
Add the ability to show an alert/status message on every page with
the themed heading. This bar will be used on test sites to indicate
that the site is for testing as well as when the site is in
maintenance mode. It also can be used to show a custom message
in a less urgent color when $alert is set in site_vars.php.