Update quizzes to match current Unicode impl
Update the quizzes and tutorial wording so it matches our current
Unicode implementation does. Of note is that the thorn quiz includes
a Unicode combined character which we do not yet support and thus fall
back to the diacritical markup.

We specifically leave the Greek transliteration quizzes in place as
some PMs will want to continue using that vs enabling the entire
polytonic glyphset.