Use a subquery to load WC stats for performance
Doing a DISTINCT join on wordcheck_events slows down performance for
very large projects with lots of rows and rounds. Using a subquery
removes the need for DISTINCT and increases performance significantly.
This is still slower (~2x) than before we pulled in the WordCheck info
but the value is worth the query cost.