Use `db_schema.sql` for table names
The database on the test server might contain some left over tables
which are no longer relevant, so instead of querying the database for
the names of all tables that currently exist in it, we have decided to
check the `db_schema.sql` file which contains all the important tables.

The only table (group of tables) that is not defined in that file are
the projectID tables which are represented by the
`projectIDxxxxxxxxxxxxx` table which should be created before running
this script.